+++ index.html
@@ -46,7 +46,6 @@
             </div>
         </div>
 
-        <script src="index.js"></script>
-        <script src="app.js" type="module"></script>
+        <script src="./js/index.js" type="module"></script>
     </body>
 </html>
